AndroidHost non-resurrectable by default, use AndroidResurrectableHost instead
